requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
plotly>=5.0.0
pypdf>=3.0.0
python-docx>=1.0.0
Pillow>=10.0.0
